update api url

+++ src/services/locationActivities_Service/location.activity.service.ts
@@ -1,4 +1,5 @@
 import { PrismaClient } from "@prisma/client";
+
 class Location_Activity_Service {
   private prisma = new PrismaClient();
   constructor() {
@@ -16,7 +17,7 @@ class Location_Activity_Service {
       } );
       return {
         status: "Success !",
-        statusCode: 201,
+        statusCode: 200,
         data: allData,
         page,
         pageSize,
@@ -29,54 +30,91 @@ class Location_Activity_Service {
       await this.prisma.$disconnect();
     }
   };
-  getLoctionActivityByName = async (nameOfActivity: string) => {
+  getLoctionActivityByName = async (activityName: string) => {
     try {
       await this.prisma.$connect();
-      const data = await this.prisma.location_Activity.findFirst({
+      const data = await this.prisma.location_Activity.findMany({
         where: {
-          activityName: nameOfActivity,
+          activityName: activityName,
         },
       });
-      return data;
+      return  data && data.length >0 ?{
+        status:'Success',
+        statusCode:200,
+        data:data
+      }:{  status:'Not found!',
+            statusCode:404}
+       
+
+  
     } catch (error) {
       console.log(error);
     } finally {
       await this.prisma.$disconnect();
     }
   };
-  // getActivityById =async (activity_id :number)=>{
-  //     try{
-  //         await this.prisma.$connect()
-  //         const data = await this.prisma.location_Activity.findUnique({
-  //             where :{
-  //                 id:activity_id
-  //             }
-  //         })
-  //         if(data){
-  //             return {
-  //                 status:'SUCCESS',
-  //                 satusCode:203,
-  //                 data:data
-  //             }
-  //         }
-  //         return  {
-  //                 status :'Error',
-  //                 statusCode :400,
-  //                 Message: 'get stuck'
-  //         }
+  getByLocationId =async(locationId:number)=>{
+    try{
+      await this.prisma.$connect
+      const rs =await this.prisma.location_Activity.findMany({
+        where:{
+          locationId:locationId
+        }
+      })
+     
+      return  rs && rs.length >0 ? {
+        status:'Success',
+        statusCode:200,
+        data:rs
 
-  //     }catch(error){
-  //         return {
-  //             status :'Error',
-  //             statusCode :500,
-  //             Message: 'Internal Internal Server'
+      }:{
+        status:'Not found',
+        statusCode:404,
 
-  //         }
+      }
 
-  //     }finally{
-  //         await this.prisma.$disconnect()
-  //     }
-  // }
+    }catch(error){
+      return {
+        status:'Internal server',
+        statusCode:500,
+        EM:error
+      }
+    }
+  }
+  getActivityById =async (activity_id :number)=>{
+      try{
+          await this.prisma.$connect()
+          const data = await this.prisma.location_Activity.findUnique({
+              where :{
+                  id:activity_id
+              }
+          })
+          if(data){
+              return {
+                  status:'SUCCESS',
+                  satusCode:200,
+                  data:data
+              }
+          }else {
+             return {status:'Not found !',
+                     statusCod:404
+            
+            }
+          }
+         
+
+      }catch(error){
+          return {
+              status :'Error',
+              statusCode :500,
+              Message: 'Internal Internal Server'
+
+          }
+
+      }finally{
+          await this.prisma.$disconnect()
+      }
+  }
   deleteLocationActivityById = async (id: number) => {
     try {
       await this.prisma.$connect();
@@ -87,9 +125,8 @@ class Location_Activity_Service {
       });
 
       return {
-        StatusCode: 200,
-        status: "Success",
-        message: "Delete successfully",
+        StatusCode: 204,
+        status: "Delete successfully!"
       };
     } catch (error) {
       console.log(error);
@@ -122,7 +159,7 @@ class Location_Activity_Service {
         },
       });
       return {
-        statusCode: 203,
+        statusCode: 200,
         status: "Update successfully",
         data: data,
       };
@@ -135,13 +172,20 @@ class Location_Activity_Service {
   deleteActivity = async (id: number) => {
     try {
       await this.prisma.$connect();
+      if(id=== null|| id===undefined){
+        return {
+          statusCode: 404,
+          status: "Resource not found!",
+        };
+
+      }
       await this.prisma.location_Activity.delete({
         where: {
           id: id, // Make sure 'id' is defined and passed correctly
         },
       });
       return {
-        statusCode: 203,
+        statusCode: 204,
         status: "Delete location activity successfully!",
       };
     } catch (error) {
@@ -175,7 +219,7 @@ class Location_Activity_Service {
       });
       return {
         status: "Success",
-        statusCode: 203,
+        statusCode: 201,
         createNewActivity,
       };
     } catch (error) {
